move src/flush.rs -> src/io/flush.rs
